Replace some functions from `app.dart` to `utils.dart`
